Feat : 구매, 판매 별 UI 정리
