+++ src/component/common/Layout/Layout.tsx
@@ -1,11 +1,21 @@
 import React from 'react';
-import { Outlet, useLocation } from 'react-router-dom';
+import { Outlet, useLocation, useParams } from 'react-router-dom';
 import Navigation from '../Navigation/Navigation';
 
 const Layout = () => {
 	const location = useLocation();
 	const currentPath = location.pathname;
+	const { productId } = useParams();
+
+	console.log(productId);
+
 	const navPages = ['/main', '/location', '/', '/purchase', '/sales'];
+
+	if (productId) {
+		navPages.push(`/purchase/detail/${productId}`);
+		navPages.push(`/sales/detail/${productId}`);
+	}
+
 	const isNavPage = navPages.includes(currentPath);
 	return (
 		<div className="bg-borderGray min-h-screen relative">
