+++ src/pages/mypage/component/salesHistory/Sold.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import BuyProduct from './BuyProduct';
-import SoldDetail from './SoldDetail';
 import { useLocation, Outlet } from 'react-router-dom';
+import SoldListProd from './SoldListProd';
+
 const Sold = () => {
 	const location = useLocation();
 	const currentPath = location.pathname;
 
 	// 판매완료
-	return <div>{currentPath === '/sales' ? <BuyProduct /> : <Outlet />}</div>;
+	return <div>{currentPath === '/sales' ? <SoldListProd /> : <Outlet />}</div>;
 };
 
 export default Sold;
